+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed MechanicalSnecko: costs 3(2) instead of 2(1)
 * Changed FragmentationGrenade: shuffles random trinkets
 * Changed GatherParts: shuffles chosen trinkets
+* Renamed Assemble to EnergyBarrier
 
 ### v0.1.0
 * Initial release